% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{combine_calendar}
\alias{combine_calendar}
\title{Get all valid service_id and operating day combinations.}
\usage{
combine_calendar(calendar, calendar_dates)
}
\value{
A tibble with columns service_id and date.
}
\description{
Get all valid service_id and operating day combinations.
}
