% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{get_service_ids}
\alias{get_service_ids}
\title{Get all valid service_ids for a given operating day.}
\usage{
get_service_ids(calendar, calendar_dates = NULL, date)
}
\description{
Get all valid service_ids for a given operating day.
}
