% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{get_trip_calendar}
\alias{get_trip_calendar}
\title{Get all valid trip_id and operating day combinations.}
\usage{
get_trip_calendar(trips, calendar, calendar_dates = NULL)
}
\value{
A tibble with columns trip_id and date.
}
\description{
Get all valid trip_id and operating day combinations.
}
