% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{prepare_calendar}
\alias{prepare_calendar}
\title{Read calendar and get valid service_id and operating day combinations.}
\usage{
prepare_calendar(calendar_csv_filename, calendar_dates_csv_filename)
}
\value{
A tibble with columns service_id and date.
}
\description{
Read calendar and get valid service_id and operating day combinations.
}
